+++ test/acceptance/features/details/page-objects/Details.js
@@ -1,4 +1,4 @@
-const { getSelectorForElementWithText } = require('../../../common/selectors')
+const { getSelectorForElementWithText } = require('../../../helpers/selectors')
 
 module.exports = {
   elements: {
